requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

# azure-functions
# azure-functions-wsgi-adapter

azure-functions==1.7.2
azure-functions-wsgi-adapter==0.2
psycopg2==2.8.6
psycopg2-binary==2.9.1
decorator==4.4.2
Django==3.2.6
django-model-utils==4.1.1
django-cors-headers==3.7.0
djangorestframework==3.12.2
sqlparse==0.4.1
pytz==2021.1
typing-extensions==3.10.0.0
asgiref==3.4.1
whitenoise==5.3.0
# asgiref==3.4.1
# Django==3.2.6
# pytz==2021.1
# sqlparse==0.4.1
# typing-extensions==3.10.0.0
# six==1.15.0

# azure-functions
# azure-functions-wsgi-adapter
# psycopg2
# psycopg2-binary
# decorator
# Django
# django-model-utils
# django-cors-headers
# djangorestframework
# sqlparse
# pytz

# cffi==1.14.5
# channels==3.0.3
# channels-redis==3.2.0
# click==7.1.2
# click-didyoumean==0.0.3
# click-plugins==1.1.1
# click-repl==0.1.6
# colormath==3.0.0
# constantly==15.1.0
# cryptography==3.4.6
# cycler==0.10.0
# daphne==3.0.1
# decorator==4.4.2
# Django==3.1.6
# django-cors-headers==3.7.0
# django-model-utils==4.1.1
# django-redis==4.12.1
# django-storages==1.11.1
# django-webpack-loader==1.0.0
# djangorestframework==3.12.2
# ExifRead==2.3.2
# gast==0.2.2
# google-pasta==0.2.0
# grpcio==1.37.1
# h5py==2.10.0
# hiredis==1.1.0
# hyperlink==21.0.0
# idna==3.1
# imageio==2.9.0
# importlib-metadata==4.0.1
# incremental==21.3.0
# jmespath==0.10.0
# joblib==1.0.1
# jsonfield==3.1.0
# Keras==2.2.5
# Keras-Applications==1.0.8
# Keras-Preprocessing==1.1.2
# kiwisolver==1.3.1
# kombu==5.0.2
# Markdown==3.3.4
# matplotlib==3.4.1
# msgpack==1.0.2
# mypy-extensions==0.4.3
# networkx==2.5
# numpy==1.20.0
# opencv-contrib-python==4.5.2.52
# opencv-python==4.5.2.52
# opt-einsum==3.3.0
# pathspec==0.8.1
# Pillow==8.1.0
# prompt-toolkit==3.0.18
# protobuf==3.16.0
# psycopg2==2.7.7
# pyasn1==0.4.8
# pyasn1-modules==0.2.8
# pycparser==2.20
# pyOpenSSL==20.0.1
# pyparsing==2.4.7
# python-dateutil==2.8.1

# PyWavelets==1.1.1
# PyYAML==5.4.1
# rawkit==0.6.0
# rawpy==0.16.0
# redis==3.5.3
# regex==2021.4.4
# s3transfer==0.3.4
# scikit-image==0.18.1
# scikit-learn==0.24.2
# scipy==1.6.3
# service-identity==18.1.0
# six==1.15.0

# swapper==1.1.2.post1
# tensorboard==1.15.0
# tensorflow==1.15.0
# tensorflow-estimator==1.15.1
# tensorflow-gpu-estimator==2.1.0
# termcolor==1.1.0
# threadpoolctl==2.2.0
# tifffile==2021.4.8
# toml==0.10.2
# Twisted==21.2.0
# txaio==21.2.1
# typed-ast==1.4.3

# urllib3==1.26.3
